BUG#20743468: ASSERTION `OLD_VALUE >= 1' FAILED. | ABORT (SIG=6) IN GTID_STATE::END_ANONYMOUS_ BUG#20748502: ASSERTION `THD->VARIABLES.GTID_NEXT.TYPE== ANONYMOUS_GROUP' FAILED.
With the binary log turned off, the server, due to a combination of
statements (which included a BINLOG statement executing an
Format_description_log_event), would hit an assertion. The reason is
that the server would incorrectly run the routine to handle the end of
a gtid violating transaction. Incorrectly, since the binary log is
disabled, the transaction should not have been handled as an offending
transaction.

In a nutshell:

- One transaction got flagged as GTID-violation, even though it
  should not since the binary log was disabled. The flag was not
  cleared at the end of the transaction;

- BINLOG statement changed GTID_NEXT to ANONYMOUS for the next
  transaction to be executed;

- The second transaction had the GTID-violation flag set and
  GTID_NEXT set to anonymous when it ended. This triggered an assertion
  when decreasing the counter of GTID-violation transactions.

This patch extends the condition that makes the
is_ddl/dml_gtid_compatible skip the compatibility testing. It now
checks if the binary log is open or not (in addition to checking if
the user has temporarily disabled it). If it not open, then the
compatibility checks are skipped. This makes the server not activate
flags that would then trigger the incorrect run of the routine to
handle gtid violating transactions, ultimately resulting in an
assertion.

Added test cases.

This patch also fixes the assert observed on BUG#20748502 .