Bug# 19823076 : READ OF FREED MEMORY IN MY_MB_WC_SJIS WITH                 SOUNDS LIKE OPERATOR IN SUBQUERY
ISSUE:
------
A correlated subquery with a having clause in the inner
query, that refers to a column in the select-list of the
outer query should return an error.

SOLUTION:
---------

The column in the inner query is incorrectly resolved to
a column in the select-list. This happens because
Item_ref::fix_fields and Item_field::fix_outer_field looks
to resolve the column in the outer context's select-list.
In this case, since the inner query is located in the outer
query's where-clause, the resolution should happen in the
from-list. The parsing_place flag needs to be taken into
account before an attempt at resolution is made.

Alias from select list can be referred from only having,
group by and order by.