Bug#17832047: Crash in calculate_materialization_costs
The problem occurs because we attempt to access an array inside
the executor using a wrong index. The index is calculated based
on table number, but the table number is incorrect after a
semi-join materialization. This is due to the column in question
being wrapped inside an Item_cache object, which does not have
functionality for Item::fix_after_pullout().

The fix is to create an implementation of Item::fix_after_pullout()
for class Item_cache that goes down into member "example" if it is
non-null.