Bug#21074643: SERVER SETS OPEN_FILES_LIMIT UNCONDITIONALLY
The problem was that when open_files_limit was set as part of
server startup, the non-Windows check for if this value made
sense didn't work. The problem is that the check only worked
of RLIMIT_NOFILE was defined while not including the system
header (sys/resource.h) that actually defines the symbol.
This made it possible to e.g. set the value to 1000000 even
on a system where the open files OS limit was 1024.

This is a 5.7 regression. On 5.6 the header file was
included indirectly via other header files.

This patch fixes the problem by including the missing header
file + removing the check on if the symbol is defined so
that similar regressions in the future will give
compilation error.