Bug#20972309 ASSERT FAILURE WHILE INITIALIZING THE DATABASE WITH INNODB_PAGE_SIZE=64K
Analysis:
We require block1 or block2 in doublewrite buffer should be in the same
data file, but in the test, block2 crosses two data files.

Solution:
Enforce minimum size of first data file to be TRX_SYS_DOUBLEWRITE_BLOCK_SIZE * 3,
so that the doublewrite buffer blocks are in the first data file.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 8916