Bug#21061746 ASSERTION `GET_WKB_GEOTYPE(P + 1) == GEOMETRY::WKB_LINESTRING' FAILED
Issue:

In BG_geometry_collection::merge_one_run(), the wkbres String buffer was
overwritten because the buffer was used across multiple executions of the
loop and it's state made wrong by such reuse.

Fix:

Define the wkbres String object in the scope where it's used so that it always
has a clean state and the memory it holds won't be overwritten.