MCP_BUG20701918  create-old-temporals MySQLD option
This patch adds the --create-old-temporals option to MySQLD allowing it
to optionally create original temporal types when temporal type
specifications have no explicit fractional seconds part (fsp).

This in turn means that the Binlog and Storage Engine types used to
store the temporal types can be compatible, easing upgrade to MySQL 5.6
based releases, especially of MySQL Cluster, and systems with
replication in use.

This is an MCP (MySQL Cluster Patch), as there is not currently
any agreement to have this included in mainline 5.6 or later
releases.

Related testcases :
 - parser_old_temporals
 - ndb_wl946*
 - ndb_binlog_temporals*