Bug #20229614 : OR CONDITIONS ON MULTI-COLUMN INDEX MAY NOT                 USE ALL INDEX COLUMNS TO FILTER ROWS
ISSUE:
------
While optimizing a range scan for the OR-operator, key_or
incorrectly assumes an out-of-memory situation. This in turn
results in an incomplete condition being considered for the
query plan and gives an incorrect row estimate. Hence the
over-estimation of the number of qualifying rows.


SOLUTION:
---------
In key_or(), a key condition is cloned if it's already in
use in some other condition. In this case NULL should be
returned only if allocation fails.

The check of key_count > 0 was originally added as part of
Bug #4157. This was done to check the key_count after the
swap of keys. Here changing the "||" to "&&" ensures that
when the new key_count is greater than 0, there will be
an attempt to clone the tree.