Introduce a mimicry of std::atomic and use it
Instead of directly reading and assigning to some variables on whose
reads/writes we depend to be atomic, introduce a mimicry of std::atomic
class and use it instead.

This way it is clear which reads and writes must be atomic. And also it
will be trivial to switch to C++11's std::atomic once we begin compiling
in C++11 mode.

WL#7170 InnoDB buffer estimates for tables and indexes