Bug#20982756: Crash in Table_list::fetch_number_of_rows
The root problem here is that for a SET command with subqueries,
st_select_lex::apply_local_transforms() is not called after resolving
the subqueries. Because of this, "first_execution" is not cleared,
and a materialized table is never set up for execution, see
st_select_lex::resolve_derived() for details.

The fix is thus to call this function for SET commands with subqueries.
To cover all possible uses of subqueries within SET, this is done
by extending the test for when to call apply_local_transforms()
inside st_select_lex::prepare().

The reason why we have a materialized table in the first place also
requires some comments: There are two views involved, and in a
SELECT statement they would both be merged into the query.

We will first try to merge table v3 into v2. But since
LEX::can_use_merged() returns false for a SET command, and
the outer query (the view v2) is not a subquery, the view is
materialized and not merged. When attempting to merge table v2 into
the outer query, LEX::can_use_merged() still returns false, but
the view is referenced in a subquery, so this view is actually merged.

This practice is a bit strange: SET commands apparently do not
allow view merging, but since the view is contained in a subquery,
the merge is allowed anyway. But the view inside view is not allowed
to be merged, even though it would be merged into the same subquery.

Thus, we seem to have two additional problems:

1. View merging should generally be allowed for SET commands, and

2. LEX::can_use_merged() may erroneously cause views contained in
   views to not be merged.

None of these problems are handled in the current bugfix.

sp.test was changed because some errors are now discovered in the
prepare phase instead of in the execute phase.