Bug #16621582: REPLICATION FAILS WITH ERROR 1837 WHEN DOING MULTI-TABLE UPDATES WITH SAVEPOINTS
A flag that stated that transactional tables were
involved in the statements was not being set. The
binlogging routing would then assume that it had
to write to the statement cache instead of the
transactional cache. This would split the transaction
in two and the savepoints would be logged without
a GTID.

This patch fixes the issue by setting the flag even
if no rows had been changed by the statement.

Added test case.