Bug #21086723	PAGE_ZIP_VERIFY_CHECKSUM() RETURNS FALSE FOR A 		VALID COMPRESSED PAGE
PROBLEM

While copying the dirty page from uncompressed page to a compressed page
we call page_zip_verify_checksum() before setting the LSN of the
compressed page,because of this LSN of compressed page is zero and if
the calculated checksum also is zero (which can happen for a valid page)
we think it is empty page and assert.

FIX

Move page_zip_verify_checksum() after setting the LSN of compressed page.
Since checksum algorithm will not consider the change in header while
calculating the checksum,this will not effect the checksum calculation.