bug#20095208 - UNABLE TO STORE FRAGMENT DURING LCP. NDBFS ERROR: 2812 - DBLQH (LINE: 14305)
On some branches we have seen very rare cases where NDBFS
FSAPPENDREQ gets an errno which only applies to sockets.
A theory is that a transporter thread has closed a file fd
and re-opened it as a socket fd.

This patch adds debug, also included in release builds.

- add filetype and fd to FSOPENCONF for signal trace

- check in PosixAsyncFile that filetype has not changed
in append and close requests

- check in ndb_socket_posix when closing a valid fd
that the fd is a socket

(copied from commit 33fe7811cf39fa237e903f3cf4b312cc7cd27fd5)