Bug#20712720 CRASH IN _ZN22REPLICATION_THREAD_API18INITIALIZE_CHANNEL
Whenever a channel initialization was attempted using the new channel
service interface and the server had its server-id set to 0 a crash
would happen. This is because the server had not properly initialized
some replication structures.

To solve this, the existing method for the initialization of the
interface will now check the server-id and it improper return an error
code. We also check if the relay log and master info respositories are
set to table, another required condition for the creation of channels.
This method shall then be used to check if the channel service
interface can be used or not.