Bug#20665051 SQL_SHOW.CC:7764: ASSERTION `QEP_TAB->CONDITION() == QEP_TAB->CONDITION_OPTIM()
The scalar subquery is uncorrelated, so it's evaluated during top
JOIN's optimize(). This evaluation (JOIN::exec) calls JOIN::prepare_result()
which materializes the I_S content into a tmp table then sorts this table.
Then EXPLAIN explains the subquery: it calls JOIN::prepare_result() a
second time, causing again materialization and sorting.
This is inefficient, and causes an assertion failure because the
I_S materialization expects to run before sorting (so it crashes because
of the previous sorting).

Fix: don't prepare result if already done.