Fixed bug#20454979: CRASH IN ACTUAL_KEY_PARTS
Because the index on hash_field is not unique, check_unique_constraint
function helps to eliminate the duplications. During such an elimination, the
index on hash_field is always needed.

In order to get the count of rows with blob fields, server will call
ha_records(). This function requires to close all indexes before calling it.
So we have to re-initialize the index on hash_field after this call.