Skip ibuf and temp-table pages in stats accounting
Move the check whether a page is an ibuf page or belongs to a temporary
table to a private method inside buf_stat_per_index_t to avoid
duplicating that check in all places where inc() or dec() methods are
called.

Move the buf_stat_per_index_t class into a dedicated header buf0stats.h
to avoid circular header dependencies. It now needs to include
ibuf0ibuf.h and fsp0sysspace.h in order to check of a page belongs to
ibuf or to a temporary table.

Assert that index id can fit into 32 bits. ibuf index id does not fulful
this because it has the magic constant id of 0xFFFFFFFF00000000.

WL#7170 InnoDB buffer estimates for tables and indexes