Fixed bug#20757211: GENERATED COLUMNS: ALTER TABLE CRASHES IN FIND_FIELD_IN_TABLE
Block subselect in generated expression.

In current wl411, the generated expression is just do a parse and doestn't
open any table. If an subquery exists in generated expression, the tables used
for the subquery can't be opened, which can result in such an error like bug
reported. Before this fix, wl411 is trying to use has_subquery() to block
subquery during fix_fields_gcol_func. However assertions might be triggered
before this block because of the unopened table.

In order to fix such a bug, we have to block the subquery before
fix_fields_gcol_func. Moreover, we don't need has_subquery() any more after we
block the subuqery at the early time.