BUG#20980271: FIX GTID MODE TO WORK CORRECTLY IN --BOOTSTRAP/--INITIALIZE MODE
Patch to enable the needed GTID infrastructure when in
bootstrap mode.

The server was not starting the necessary memory structures
to setup GTIDs when on bootstrap mode. This patch makes the
server generate the auto.cnf file and set up GTID sets and
the sidmap when operating in bootstrap mode.

It also changes MTR. It needs to because the bootstrap of the
server is done once, and then the same install dir is used
for initializing the master and slave. Thence, we need to
remove the auto.cnf file (which holds the UUID generated
UUID at bootstrap time), before copying the install dir for
provisioning the new server that MTR spawns. This was not
needed before, because mysqld --initialize would not
setup the UUID for the server, i.e., the auto.cnf file.