BUG#20171740 - OVERRIDE DEFAULT SERVICE START/STOP                TIMEOUT USED BY SYSTEMD FOR MYSQL.
Server startup and shutdown time is not constant
and bounded and may take several minutes. This
is because of InnoDB which does disk I/O operations
that depend on InnoDB log file size, preload and dump
of dump buffer pools, purge threads being active during
shutdown etc. The default service start and stop timeout
specified by systemd has a small value and this triggers
the restart logic of systemd causing another instance of
the mysqld to be spawned.

The fix is to disable the service start/stop timeout logic of
systemd by specifying TimeoutSec=0 in the service file for
mysqld service. Also fix the comment in mysys_err.h.