BUG#20593028 MORE CONCURRENCY IN SHOW SLAVE STATUS: MAKE              LOCK_MSR_MAP AN RWLOCK
This patch changed the replication channel map lock to a
Checkable_rwlock to reduce the contention while searching for
master_info objects based on replication channel names.

But in order to be conservative, we are always taking the write
lock.

The lock was moved to the channel_map object and it is not global
anymore.