BUG#20221956 LOCK_MSR_MAP IS NOT ASSERTED IN MULTISOURCE_INFO METHODS              DEALING WITH MSR_MAP
According to current code documentation, the Multisource_info class "is
not yet thread safe. Any part of replication code that calls this class
member functions should always lock the channel_map lock."

But there is no asserts for the lock on Multisource_info class methods.

This patch adds asserts for channel_map lock on functions that deal
with it.