Bug#20914786 : MYSQL_UPGRADE FAILS WITH SHOW_COMPATIBILITY_56 = OFF
This is a rework on the previous fix for mysl_upgrade.

mysql_upgrade can -- not -- use performance_schema tables
to read the server version,
or use SHOW VARIABLES that rely on the same tables,
because these tables might not even be installed (by the upgrade script)
yet.

Use
  SELECT @@global.version
  SELECT @@global.datadir
which is robust and also works for in place upgrades.