Bug#17846865 HANDLER READING INDEX ON GEOMETRY COLUMN CRASH
Issue:

In the test case query, we are scanning an btree index built upon a
geometry column, but target value is 1. When trying to store 1 into
the geometry field of the record buffer, there is a type conversion
error because integers can't be used as a GEOMETRY byte string, and the
record buffer's geometry field has nothing stored in.

However we didn't stop the index search in case of this error condition,
but always tried to fetch the target row by creating the index key
to search, and then search the index for it. And in this case the blob
pointer in the field can be any value, so we got an memory access
violation when building the target index key.

Fix:

Do not search index if target index value is not stored into row buffer
successfully.