Bug#21153716 ST_CONVEXHULL RETURNS WRONG RESULT
Issue:

The colinear check algorithm requires the collection of points given are
unique, otherwise the result may or may not be correct. In this case
duplicate points are given and they happen to reveal the issue.

Fix:

When seeing if a collection of points are colinear, make them unique first.