Fixed bug#21142905: PARTITIONED GENERATED COLS - !TABLE || (!TABLE->WRITE_SET                                                         || BITMAP_IS_SET
This bug is caused by not marking virtual generated columns to be writeable
when it tries to read it.

If a partitioned column is writeable, the partition might be changed.
Therefore, server marks all columns readable so that the row can be moved to
the right partition. During setting all columns readable, virtual generated
columns are not set to be writeable. Even if the partitioned column is not writable
, it's also requirted to be readable so that to calculate the partition id to
place updated or deleted records. Such a set also needs mark virtual generated columns
writable.