Bug#20834483 - ASSERTION: TABLE->SPACE == FIL_SPACE_GET_ID_BY_NAME( TABLE->TABLESPACE())
Problem:
--------
Truncate on tables with FTS indexes using General Tablespace assume the
tablespace name is always populated correctly in table cache(dict_table_t).

Fix:
----
Do not assume the table->tablespace() is always populated. We now fetch the
real name if table->tablespace() is NULL or if the name is "innodb_general"

Reviewed-by: Kevin Lewis<kevin.lewis@oracle.com>
RB: 9063