Bug #18194196: OPTIMIZER EXECUTES STATEMENT INPERFORMANT
Problem:
While choosing the join order, the cost for doing table_scan
is wrongly calculated. As a result table_scan is preferred
over eq_ref, thereby choosing a bad plan.

Analysis:

While calculating the fanout in semijoin_dupsweedout method,
if an inner table is ahead of the outer table in the join order,
fanout is not calculated correctly. This is what is happening
w.r.t the query in the bugpage.

As seen from the trace, a table scan is preferred over eq_ref.
This is because when calculating the cost for eq_ref, optimizer
takes into consideration the correct prefix_row_count calculated
in prev_records_read. In this case the first table emailstorerel
 has around 42000 records and the next one is a ref scan with 1 row as
fanout and the next one is also 1. Cost for eq_ref is calculated
based on this.

While in table_scan, cost is calculated based on row_count
passed to best_access_path which ideally should be the prefix_row_count
based on the partial plan chosen. In this case it should be more than
42000. It is currently 1. This changes to "1"  after semi-join strategy
for dups_weedout is checked after choosing emailstore as the second table
in the following part of trace.

                    "plan_prefix": [
                      "`emailstoreorel`"
                    ],
                    "table": "`emailstore`",
                    "best_access_path": {
                      "considered_access_paths": [
                        {
                          "access_type": "ref",
                          "index": "PRIMARY",
                          "rows": 1,
                          "cost": 42092,
                          "chosen": true
                        },
                        {
                          "access_type": "scan",
                          "using_join_cache": true,
                          "rows": 70554,
                          "cost": 5.94e8,
                          "chosen": false
                        }
                      ]
                    },
                    "cost_for_plan": 67583,
                    "rows_for_plan": 42092,
                    "semijoin_strategy_choice": [
                      {
                        "strategy": "DuplicatesWeedout",
                        "cost": 76004,
                        "rows": 1,
                        "duplicate_tables_left": true,
                        "chosen": true
                      }
                    ],
So after Optimize_table_order::semijoin_dupsweedout_access_paths(), the
current_rowcount now becomes "1" because the prefix_row_count is
calculated solely on outer fanout (from table emailstore) which in this case
is 1. The inner fanout from table emailstoreorel is not considered at all.

Solution:
Change the current formulas to the one's mentioned in the todo text.
The formulas now take into consideration all the scenarios which can have inner
tables ahead of outer tables in a join order. In such a scenario, if
inner_fanout is more than 1, this will be moved to outer_fanout and inner_fanout
is re-calculated.
max_outer_fanout is introduced to keep a cap on outer_fanout not to exceed the
cardinality of the cross product of outer tables.

changes to test files:
Two sets of changes can be noted.
1. When a inner table with full table scan is chosen as the first table in the join
order, earlier fanout was wrongly calculated. As a result the cost for doing
dups weedout was less. With the current formulas, the cost for doing writes become
more because of increased outer fanout in these cases. As a result the cost is not less
than materialized scan.

2. Optimizer now calculates the cost of table scan correctly. As a result, eq_ref
is preferred over table scan in these cases.